+++ src/studio/src/designer/frontend/packages/schema-model/src/index.ts
@@ -2,16 +2,22 @@ export { buildJsonSchema } from './lib/build-json-schema';
 export { buildUiSchema } from './lib/build-ui-schema';
 export { getRestrictions, StrRestrictionKeys, castRestrictionType } from './lib/restrictions';
 export type { UiSchemaNode, UiSchemaNodes } from './lib/types';
-export { ROOT_POINTER, ObjectKind, Keywords, FieldType, CombinationKind, Capabilites } from './lib/types';
-export { createNodeBase, replaceLastPointerSegment, makePointer, pointerIsDefinition } from './lib/utils';
+export { ROOT_POINTER, ObjectKind, Keywords, FieldType, CombinationKind } from './lib/types';
+export {
+  createNodeBase,
+  replaceLastPointerSegment,
+  makePointer,
+  pointerIsDefinition,
+  combinationIsNullable,
+  pointerExists,
+} from './lib/utils';
 export { createChildNode } from './lib/mutations/create-node';
 export { removeNodeByPointer } from './lib/mutations/remove-node';
 export { renameNodePointer } from './lib/mutations/rename-node';
-export { promotePropertyToType } from './lib/mutations/promote-node';
-export { getCapabilities } from './lib/capabilities';
+export { convertPropToType } from './lib/mutations/promote-node';
+export { getCapabilities, Capabilites } from './lib/capabilities';
 export { toggleArrayAndField, canToggleArrayAndField } from './lib/mutations/toggle-array-field';
 export {
-  combinationIsNullable,
   getChildNodesByNode,
   getChildNodesByPointer,
   getNodeByPointer,
@@ -19,5 +25,4 @@ export {
   getNodeIndexByPointer,
   getRootNodes,
   getUniqueNodePath,
-  pointerExists,
 } from './lib/selectors';
