+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/capabilities.test.ts
@@ -1,6 +1,6 @@
-import { getCapabilities } from './capabilities';
+import { Capabilites, getCapabilities } from './capabilities';
 import { createNodeBase } from './utils';
-import { Capabilites, FieldType, ObjectKind } from './types';
+import { FieldType, ObjectKind } from './types';
 import { expect } from '@jest/globals';
 
 test('that capabilities work as expected', () => {
