+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/mutations/promote-node.test.ts
@@ -2,12 +2,12 @@ import { buildUiSchema } from '../build-ui-schema';
 import { buildJsonSchema } from '../build-json-schema';
 import { FieldType, Keywords } from '../types';
 import { makePointer } from '../utils';
-import { promotePropertyToType } from './promote-node';
+import { convertPropToType } from './promote-node';
 import { simpleTestJsonSchema } from '../../../test/testUtils';
 
 test('that we can promote a node', () => {
   const originalNodeMap = buildUiSchema(simpleTestJsonSchema);
-  const promotedNodeMap = promotePropertyToType(originalNodeMap, makePointer(Keywords.Properties, 'world'));
+  const promotedNodeMap = convertPropToType(originalNodeMap, makePointer(Keywords.Properties, 'world'));
   expect(buildJsonSchema(promotedNodeMap)).toEqual({
     [Keywords.Properties]: {
       hello: { [Keywords.Type]: FieldType.String },
@@ -26,6 +26,6 @@ test('that promotePropertyToType throws errors', () => {
       email: { [Keywords.Type]: FieldType.String },
     },
   });
-  expect(() => promotePropertyToType(uiSchemaNodes, '#/$defs/email')).toThrowError();
-  expect(() => promotePropertyToType(uiSchemaNodes, '#/properties/email')).toThrowError();
+  expect(() => convertPropToType(uiSchemaNodes, '#/$defs/email')).toThrowError();
+  expect(() => convertPropToType(uiSchemaNodes, '#/properties/email')).toThrowError();
 });
