+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/utils.test.ts
@@ -1,5 +1,5 @@
-import { createNodeBase, getParentNodeByPointer, makePointer, sortNodesByChildren } from './utils';
-import { ROOT_POINTER } from './types';
+import { combinationIsNullable, createNodeBase, getParentNodeByPointer, makePointer } from './utils';
+import { FieldType, ROOT_POINTER } from './types';
 import { getGeneralJsonSchemaForTest } from '../../test/testUtils';
 import { buildUiSchema } from './build-ui-schema';
 import { expect } from '@jest/globals';
@@ -32,18 +32,10 @@ test('that we can makePointer', () => {
   expect(makePointer('#/properties', 'hello')).toBe('#/properties/hello');
 });
 
-test('that we can sortNodes', () => {
-  const rootNode = createNodeBase(ROOT_POINTER);
-  const child1 = createNodeBase('someother path');
-  const child3 = createNodeBase('somepath');
-  rootNode.children.push(child1.pointer);
-  rootNode.children.push(child3.pointer);
-  const child2 = createNodeBase('someother path', 'another one here');
-  child1.children.push(child2.pointer);
-  const testArray = [child1, child2, child3, rootNode];
-  for (let i = 0; i < 5; i++) {
-    testArray.sort(() => Math.random() - 0.5);
-    const sorted = sortNodesByChildren(testArray);
-    expect(sorted).toStrictEqual([rootNode, child1, child2, child3]);
-  }
+test('that we can check if combination is nullable', () => {
+  const regularChild = createNodeBase('regular child');
+  const nullableChild = createNodeBase('nullable child');
+  nullableChild.fieldType = FieldType.Null;
+  expect(combinationIsNullable([regularChild, regularChild])).toBeFalsy();
+  expect(combinationIsNullable([regularChild, nullableChild])).toBeTruthy();
 });
