+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/utils.ts
@@ -41,10 +41,10 @@ export const getObjectKind = (schemaNode: JsonSchemaNode): ObjectKind => {
   }
 };
 
-export const schemaTypeIncludes = (schemaNodeType: string | string[], type: FieldType) =>
+export const schemaTypeIncludes = (schemaNodeType: string | string[], type: FieldType): boolean =>
   schemaNodeType === type || (Array.isArray(schemaNodeType) && schemaNodeType.includes(type));
 
-export const schemaTypeIsNillable = (schemaNodeType: string | string[]) =>
+export const schemaTypeIsNillable = (schemaNodeType: string | string[]): boolean =>
   schemaNodeType !== FieldType.Null && schemaTypeIncludes(schemaNodeType, FieldType.Null);
 
 export const getParentNodeByPointer = (uiSchemaNodes: UiSchemaNodes, pointer: string): UiSchemaNode | undefined => {
@@ -96,20 +96,8 @@ export const pointerReplacer = (node: UiSchemaNode, oldPointer: string, newPoint
   return nodeCopy;
 };
 
-export const sortNodesByChildren = (uiSchemaNodes: UiSchemaNodes): UiSchemaNodes => {
-  const tempMap = new Map();
-  uiSchemaNodes.forEach((node) => tempMap.set(node.pointer, node));
-  return treeWalker(tempMap, ROOT_POINTER);
-};
+export const pointerExists = (uiSchemaNodes: UiSchemaNodes, pointer: string): boolean =>
+  getNodeIndexByPointer(uiSchemaNodes, pointer) !== undefined;
 
-const treeWalker = (map: Map<string, UiSchemaNode>, pointer: string): UiSchemaNodes => {
-  const nodes = [];
-  if (map.has(pointer)) {
-    const currentNode = map.get(pointer) as UiSchemaNode;
-    nodes.push(currentNode);
-    currentNode.children.forEach((childPointer) => nodes.push(...treeWalker(map, childPointer)));
-  } else {
-    throw new Error(`Can't find ${pointer} in node-array.`);
-  }
-  return nodes;
-};
+export const combinationIsNullable = (childNodes: UiSchemaNode[]): boolean =>
+  childNodes.some((child) => child.fieldType === FieldType.Null);
