+++ src/studio/src/designer/frontend/packages/schema-model/test/testUtils.ts
@@ -80,7 +80,7 @@ export const simpleTestJsonSchema = {
     },
     world: {
       [Keywords.Properties]: {
-        hello: {
+        hola: {
           [Keywords.Type]: FieldType.Boolean,
         },
       },
